+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Next Release] - Release Date
+### Added
+- [server] When TERESA_SERVER_DEBUG is set log the stack trace on
+  panic/recover.
+
 ## [0.6.0] - 2017-08-23
 ### Added
 - [client] App name length validation