+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Next Release] - Release Date
+### Added
+- [server] Add --debug flag. For now only print the stack trace on panic/recover.
+
 ## [0.6.0] - 2017-08-23
 ### Added
 - [client] App name length validation