+++ CHANGELOG.md
@@ -9,6 +9,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - [client] Commands 'env-set' and 'env-unset' show the current cluster.
 
+### Fixed
+- [server] Restart count on 'app info' output
+
 ## [0.6.0] - 2017-08-23
 ### Added
 - [client] App name length validation